Исправляет отрисовку элементов
